Video Description- https://drive.google.com/open?id=0ByZlAUHpQ9_mUldGX3ViNEFnaHc
Working of VU meter Circuit is simple; at first we provide analog signal to pin no.

5 with the help of aux wire and female aux port. So for a higher sound we will

have higher value and lower value for a lower sound. Then these voltages signals

are fed to LM3914, which works as a voltmeter and glows LEDs according to the

intensity of sound. Pin no. 4 is used to set reference low (which is generally

grounded) and pin no. 6 is used for setting reference high. The voltage to pin no. 6

is provided by pin no. 7(ref out (reference voltage source 1.25V)) and a voltage

divider circuit. Pin no. 8 is use to adjust the reference out voltage of pin no. 7.